requirements.txt
# FastAPI Backend Dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
nest-asyncio>=1.5.8

# MCP (Model Context Protocol)
mcp>=1.0.0

# AI/ML Dependencies
anthropic>=0.7.8
pinecone>=7.3.0
tavily-python>=0.3.3
tiktoken>=0.5.0

# Document Processing
PyPDF2>=3.0.1
python-docx>=1.1.0
python-dotenv>=1.0.0

# File Generation
reportlab>=4.0.7
fpdf2>=2.7.6
markdown>=3.5.1
jinja2>=3.1.2

# HTTP Client
httpx>=0.25.2
requests>=2.31.0

# Utilities
pydantic>=2.5.0
websockets>=12.0
openai>=1.0.0

# Scientific Computing
arxiv>=2.1.0